Fixing More Issues with Responsiveness
headshot-cropped.JPG would be extremely large on some computers, switched back to grid in order to allow for a fixed size for the picture within the pic-div. I’m also converting the font sizes for all the text content relative to either the root element, or its parent element to allow for a responsive webpage. I first tinkered with the #myname div, then adjusted its child elements. I then went on to do the same for the other elements, like the #about-me text, #intro-text, and #bottom-div contents