Responsiveness updates on all webpages
Removed more portions of the usage of 'px' on the portfolio css (which gallery.html and expectations.html links to'. Also fixed responsiveness for gallery.html where appearance on desktop would be manageable, but visibility on other devices (notably mobile) would be horrid. Changed the min-width and max-height to accomodate various devices, and also increased the row sizes of the grid layout for 'first-image' to provide ample room for both the text and the photo.